Eliminate `CodlDecodable` in favor of `Decodable in Codl`
