+++ lib/cellulose/src/test/cellulose.Tests.scala
@@ -34,6 +34,7 @@ package cellulose
 
 import anticipation.*
 import contingency.*
+import distillate.*
 import eucalyptus.*, logging.stdout
 import fulminate.*
 import gossamer.*
@@ -1047,7 +1048,7 @@ object Tests extends Suite(m"Cellulose tests (Part 1)"):
         read(t"root one  two three  four five\n").wiped
       .assert(_ == CodlDoc(CodlNode(t"root")(CodlNode(t"one")(), CodlNode(t"two three")(), CodlNode(t"four five")())))
 
-    def roundtrip[T: {Encodable in Codl, CodlDecodable, CodlSchematic}](value: T): T = value.codl.as[T]
+    def roundtrip[T: {Encodable in Codl, Decodable in Codl, CodlSchematic}](value: T): T = value.codl.as[T]
 
     Tests2()
 
@@ -1056,7 +1057,7 @@ object Tests2 extends Suite(m"Cellulose tests (Part 2)"):
   given Realm = realm"tests"
 
   def run(): Unit = supervise:
-    def roundtrip[T: {Encodable in Codl, CodlDecodable, CodlSchematic}](value: T): T = value.codl.as[T]
+    def roundtrip[T: {Encodable in Codl, Decodable in Codl, CodlSchematic}](value: T): T = value.codl.as[T]
     def read(text: Text): CodlDoc = Codl.parse(text)
     suite(m"Generic Derivation tests"):
 
@@ -1138,7 +1139,7 @@ object Tests2 extends Suite(m"Cellulose tests (Part 2)"):
       val complex = Bar(List(Baz(t"a", 2, Unset), Baz(t"c", 6, 'e')), Quux(t"e", List(1, 2, 4)))
 
 
-      def print[T: {CodlDecodable, Encodable in Codl, CodlSchematic}](value: T): Text =
+      def print[T: {Decodable in Codl, Encodable in Codl, CodlSchematic}](value: T): Text =
         val writer = new ji.StringWriter()
         Printer.print(writer, value.codl)
         writer.toString().show
